separated loadImage from renderToCanvas
