+++ packages/core/src/RenderingEngine/helpers/index.ts
@@ -1,7 +1,6 @@
 import createVolumeActor from './createVolumeActor';
 import createVolumeMapper from './createVolumeMapper';
 import getOrCreateCanvas from './getOrCreateCanvas';
-import renderToCanvas from './renderToCanvas';
 import setVolumesForViewports from './setVolumesForViewports';
 import addVolumesToViewports from './addVolumesToViewports';
 
@@ -9,7 +8,6 @@ export {
   createVolumeActor,
   createVolumeMapper,
   getOrCreateCanvas,
-  renderToCanvas,
   setVolumesForViewports,
   addVolumesToViewports,
 };
