+++ packages/core/src/utilities/index.ts
@@ -19,6 +19,8 @@ import getVolumeViewportsContainingSameVolumes from './getVolumeViewportsContain
 import getVolumeViewportsContainingVolumeId from './getVolumeViewportsContainingVolumeId';
 import transformWorldToIndex from './transformWorldToIndex';
 import prefetchStack from './prefetchStack';
+import loadImageToCanvas from './loadImageToCanvas';
+import renderToCanvas from './renderToCanvas';
 
 // name spaces
 import * as planar from './planar';
@@ -48,4 +50,6 @@ export {
   getVolumeViewportsContainingVolumeId,
   transformWorldToIndex,
   prefetchStack,
+  loadImageToCanvas,
+  renderToCanvas,
 };
